move constants to lib folder
