+++ app/containers/message/Encrypted.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 
 import Touchable from './Touchable';
-import { E2E_MESSAGE_TYPE } from '../../lib/encryption/constants';
+import { E2E_MESSAGE_TYPE } from '../../lib/constants';
 import { CustomIcon } from '../../lib/Icons';
 import { themes } from '../../constants/colors';
 import { BUTTON_HIT_SLOP } from './utils';
