+++ app/lib/methods/logout.ts
@@ -9,7 +9,7 @@ import database, { getDatabase } from '../database';
 import RocketChat from '../rocketchat';
 import { useSsl } from '../../utils/url';
 import log from '../../utils/log';
-import { E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../encryption/constants';
+import { E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../constants';
 import UserPreferences from '../userPreferences';
 import { ICertificate, IRocketChat } from '../../definitions';
 import sdk from '../rocketchat/services/sdk';
