+++ app/presentation/RoomItem/LastMessage.tsx
@@ -5,7 +5,7 @@ import I18n from '../../i18n';
 import styles from './styles';
 import { MarkdownPreview } from '../../containers/markdown';
 import { themes } from '../../constants/colors';
-import { E2E_MESSAGE_TYPE, E2E_STATUS } from '../../lib/encryption/constants';
+import { E2E_MESSAGE_TYPE, E2E_STATUS } from '../../lib/constants';
 
 interface ILastMessage {
 	theme: string;
