+++ app/sagas/encryption.js
@@ -5,7 +5,7 @@ import { ENCRYPTION } from '../actions/actionsTypes';
 import { encryptionSet } from '../actions/encryption';
 import { Encryption } from '../lib/encryption';
 import Navigation from '../lib/Navigation';
-import { E2E_BANNER_TYPE, E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../lib/encryption/constants';
+import { E2E_BANNER_TYPE, E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../lib/constants';
 import database from '../lib/database';
 import RocketChat from '../lib/rocketchat';
 import UserPreferences from '../lib/userPreferences';
