+++ app/views/RoomsListView/ListHeader/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useTheme } from '../../../theme';
 import * as List from '../../../containers/List';
-import { E2E_BANNER_TYPE } from '../../../lib/encryption/constants';
+import { E2E_BANNER_TYPE } from '../../../lib/constants';
 import { themes } from '../../../constants/colors';
 import OmnichannelStatus from '../../../ee/omnichannel/containers/OmnichannelStatus';
 import { IUser } from '../../../definitions';
